Javasrcript
I learned how to incorporate some javascript which is a simple <script type="text/javascript" src="script.js"></script> in the head or body to create a simple alert sensor when ever a customer books an appointment. I used some js whenever there was some interaction with the star rating system along with a few of the buttons in the web application as well.